remove treebeard API key
hey folks -- sorry to say, due to my google cloud credits running out and a misunderstanding, my GCP account has just been nuked and uploads are out of action.

I should be able to retrieve any stored data if you need but this is likely a permanent change as I was planning on deprecating the backend.

Again, sorry for the lack of warning. Fortunately I have a more open replacement to this capability coming very soon.