removed FIXME tag
there is no need for an ```else``` statement here, because the ```if``` clause in wrapped inside a while loop with an ```else``` statement. Basically, if the ```if``` statement evaluates to ```True``` then it breaks out of the while loop, otherwise the loop counter is increased by 1. At the end, when the loop counter reaches the defined value (without the ```if``` statement ever evaluating to ```True```), then it executes the ```else``` statement, i.e. it raises a ```ValueError```