Update talktorial.ipynb
The cell IDs for "binding site detection" sections in "Tehory" and "Practical" were the same, and so "Practical" also linked to the "binding site detection" in "Theory" (because it comes first). => Changed the ID of "Binding site detection" in practical by adding an underline at the end.

Small revision to the text in "Aim of this talktorial".