fix RDKit render issue in Pandas DataFrames
This requires a newer version of RDKit, thus the pinned version was removed from `test_env.yml`; this still doesn't update RDKit to the latest version, due to pinning in other dependencies. Note that for this fix to work, you have to recreate your environment with the new `test_env.yml`, or update RDKit manually.