suppress warning encountered in cells 33, 41, 63
The warning "MDAnalysis/coordinates/PDB.py:1129: UserWarning: Found no information for attr: 'formalcharges' Using default value of '0'" is caused by MDAnalysis, which is used by OpenCADD, which is used in the `extract_molecule_from_pdb_file` function in `utils/helpers/pdb.py`. Cause of the warning is that formal charges are not present in the PDB file. This is a common situation since many PDB files don't have formal charges assigned. However, this pinned version of MDAnalysis raises a warning in these situations, which is now suppressed by a context manager in this commit