Suppress deprecation warnings in plip
plip has not updated its code with regard to new numpy versions, and thus the explicit casting `np.uint32(i)` in line 361 of `plip.basic.supplemental` (function `int32_to_negative`) throws thousands of warnings. These are suppressed here, as the warning suppression in our `util.__init__` is not catching these, probably due to setting of the logging level at the top of this module.